refactor: remove deprecated file-write-handler and update imports in related handlers to use new file writing utilities
